Add the solarized-contrast custom variable.
As requested in issue #32. I hadn’t added it previously, due to the description attached to it: “Stick with normal! It's been carefully tested. Setting this option to high or low does use the same Solarized palette but simply shifts some values up or down in order to expand or compress the tonal range displayed.”